Fix fractional sparks truncation and Grok 4 Fast token overestimate
Convert all sparks columns from INTEGER to NUMERIC(10,2) so sub-1-spark
operations (e.g. Grok 4 Fast at ~0.44 sparks) are correctly debited
instead of being truncated to 0. Remove all int() truncation in the
billing pipeline. Fix Grok image token estimator divisor (750→1600) to
match observed ~668 tokens from OpenRouter.

+++ backend/app/models/cost_decision.py
@@ -51,7 +51,7 @@ class CostDecision(Base):
 
     estimated_input_tokens: Mapped[int | None] = mapped_column(Integer, nullable=True)
     estimated_output_tokens: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    estimated_sparks: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    estimated_sparks: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
 
     cost_per_input_token: Mapped[Decimal | None] = mapped_column(
         Numeric(20, 12), nullable=True
@@ -76,7 +76,7 @@ class CostDecision(Base):
     status: Mapped[str] = mapped_column(String(16), nullable=False, default="pending")
     error_message: Mapped[str | None] = mapped_column(Text, nullable=True)
 
-    reserved_sparks: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    reserved_sparks: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
     idempotency_key: Mapped[str | None] = mapped_column(String(256), nullable=True, unique=True)
 
     created_at: Mapped[datetime] = mapped_column(
